requirements.txt
Flask
tensorflow
numpy
Pillow
openai>=1.0.0
python-dotenv
requests